+++ packages/core/src/core/orchestrator.ts
@@ -1,7 +1,7 @@
 import { Logger } from "./logger";
 import { RoomManager } from "./room-manager";
 import { TaskScheduler } from "./task-scheduler";
-import type { BaseProcessor, MessageProcessor } from "./processor";
+import type { BaseProcessor } from "./processor";
 import type { Memory, ProcessedResult, VectorDB } from "./types";
 import { HandlerRole, LogLevel, type LoggerConfig } from "./types";
 import type { IOHandler } from "./types";
@@ -238,96 +238,101 @@ export class Orchestrator {
      * and handles any follow-on "action" or "output" suggestions in a chain.
      */
     private async runAutonomousFlow(initialData: unknown, sourceName: string) {
-        // We keep a queue of "items" to process
-        const queue: Array<{ data: unknown; source: string }> = [
-            { data: initialData, source: sourceName },
-        ];
+        // A queue of "things to process"
+        const queue: Array<{ data: unknown; source: string }> = [];
 
+        // If the initial data is already an array, enqueue each item
+        if (Array.isArray(initialData)) {
+            for (const item of initialData) {
+                queue.push({ data: item, source: sourceName });
+            }
+        } else {
+            queue.push({ data: initialData, source: sourceName });
+        }
+
+        // You can keep track of any "outputs" you need to return or do something with
         const outputs: Array<{ name: string; data: any }> = [];
 
+        // Keep processing while there is something in the queue
         while (queue.length > 0) {
             const { data, source } = queue.shift()!;
 
-            // 1) Ensure there's a room
-            const room = await this.roomManager.ensureRoom(source, "core");
+            // processContent now returns an array of ProcessedResult
+            const processedResults = await this.processContent(data, source);
 
-            // 2) Find the first processor that can handle it
-            const processor = Array.from(this.processors.values()).find((p) =>
-                p.canHandle(data)
-            );
-
-            if (!processor) {
-                // this.logger.warn("No processor found for data", { source });
+            // If there's nothing to process further, continue
+            if (!processedResults || processedResults.length === 0) {
                 continue;
             }
 
-            // 3) Process with the found processor
-            const processed = await this.processContent(data, source);
-
-            // If the processor thinks we've already processed it, we skip
-            if (processed?.alreadyProcessed) {
-                continue;
-            }
-
-            // 4) Save to memory (like you do in processInputTask)
-            await this.roomManager.addMemory(
-                room.id,
-                JSON.stringify(processed?.content),
-                {
-                    source,
-                    type: "input",
-                    ...processed?.metadata,
-                    ...processed?.enrichedContext,
-                }
-            );
-
-            if (processed?.updateTasks) {
-                for (const task of processed.updateTasks) {
-                    await this.scheduleTaskInDb(
-                        task.name,
-                        task.data,
-                        task.intervalMs
-                    );
-                }
-            }
-
-            // 5) For each suggested output, see if it's an action or an output
-            for (const output of processed?.suggestedOutputs ?? []) {
-                const handler = this.ioHandlers.get(output.name);
-                if (!handler) {
-                    this.logger.warn(
-                        "No handler found for suggested output",
-                        output.name
-                    );
+            // Now handle each ProcessedResult
+            for (const processed of processedResults) {
+                // If the processor says it's already been handled, skip
+                if (processed.alreadyProcessed) {
                     continue;
                 }
 
-                if (handler.role === HandlerRole.OUTPUT) {
-                    outputs.push({ name: output.name, data: output.data });
-
-                    // Dispatch to an output handler (e.g. send a Slack message)
-                    await this.dispatchToOutput(output.name, output.data);
-                } else if (handler.role === HandlerRole.ACTION) {
-                    // Execute an action (e.g. fetch data from an API), wait for the result
-                    const actionResult = await this.dispatchToAction(
-                        output.name,
-                        output.data
-                    );
-                    // Then feed the result back into the queue, so it will be processed
-                    if (actionResult) {
-                        queue.push({
-                            data: actionResult,
-                            source: output.name, // or keep the same source, your choice
-                        });
+                // If any tasks need to be scheduled in the DB, do so
+                if (processed.updateTasks) {
+                    for (const task of processed.updateTasks) {
+                        await this.scheduleTaskInDb(
+                            task.name,
+                            task.data,
+                            task.intervalMs
+                        );
                     }
-                } else {
-                    this.logger.warn(
-                        "Suggested output has an unrecognized role",
-                        handler.role
-                    );
+                }
+
+                // For each suggested output
+                for (const output of processed.suggestedOutputs ?? []) {
+                    const handler = this.ioHandlers.get(output.name);
+                    if (!handler) {
+                        this.logger.warn(
+                            "No handler found for suggested output",
+                            output.name
+                        );
+                        continue;
+                    }
+
+                    if (handler.role === HandlerRole.OUTPUT) {
+                        // e.g. send a Slack message
+                        outputs.push({ name: output.name, data: output.data });
+                        await this.dispatchToOutput(output.name, output.data);
+                    } else if (handler.role === HandlerRole.ACTION) {
+                        // e.g. fetch data from an external API
+                        const actionResult = await this.dispatchToAction(
+                            output.name,
+                            output.data
+                        );
+                        // If the action returns new data (array or single),
+                        // feed it back into the queue to continue the flow
+                        if (actionResult) {
+                            if (Array.isArray(actionResult)) {
+                                for (const item of actionResult) {
+                                    queue.push({
+                                        data: item,
+                                        source: output.name,
+                                    });
+                                }
+                            } else {
+                                queue.push({
+                                    data: actionResult,
+                                    source: output.name,
+                                });
+                            }
+                        }
+                    } else {
+                        this.logger.warn(
+                            "Suggested output has an unrecognized role",
+                            handler.role
+                        );
+                    }
                 }
             }
         }
+
+        // If you want, you can return the final outputs array or handle it differently
+        return outputs;
     }
 
     /**
@@ -536,10 +541,51 @@ export class Orchestrator {
     public async processContent(
         content: any,
         source: string
+    ): Promise<ProcessedResult[]> {
+        // If content is already an array, process each item individually
+        if (Array.isArray(content)) {
+            const allResults: ProcessedResult[] = [];
+            for (const item of content) {
+                // (Optional) Delay to throttle
+                await new Promise((resolve) => setTimeout(resolve, 5000));
+
+                const result = await this.processContentItem(item, source);
+                if (result) {
+                    allResults.push(result);
+                }
+            }
+            return allResults;
+        }
+
+        // For single item, wrap in array
+        const singleResult = await this.processContentItem(content, source);
+        return singleResult ? [singleResult] : [];
+    }
+
+    private async processContentItem(
+        content: any,
+        source: string
     ): Promise<ProcessedResult | null> {
         let memories: Memory[] = [];
 
         if (content.room) {
+            const hasProcessed =
+                await this.roomManager.hasProcessedContentInRoom(
+                    content.contentId,
+                    content.room
+                );
+
+            if (hasProcessed) {
+                this.logger.debug(
+                    "Orchestrator.processContent",
+                    "Content already processed",
+                    {
+                        contentId: content.contentId,
+                        roomId: content.room,
+                    }
+                );
+                return null;
+            }
             const room = await this.roomManager.ensureRoom(
                 content.room,
                 source
@@ -579,10 +625,35 @@ export class Orchestrator {
             (h) => h.role === HandlerRole.ACTION
         );
 
-        return processor.process(content, JSON.stringify(memories), {
-            availableOutputs,
-            availableActions,
-        });
+        const result = await processor.process(
+            content,
+            JSON.stringify(memories),
+            {
+                availableOutputs,
+                availableActions,
+            }
+        );
+
+        if (content.room) {
+            // Save the result to memory
+            await this.roomManager.addMemory(
+                content.room,
+                JSON.stringify(result?.content),
+                {
+                    source,
+                    ...result?.metadata,
+                    ...result?.enrichedContext,
+                }
+            );
+
+            // Mark the content as processed
+            await this.roomManager.markContentAsProcessed(
+                content.contentId,
+                content.room
+            );
+        }
+
+        return result;
     }
 
     /**
